requirements.txt
Babel==2.9.1
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.8.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
openai==1.3.6
pip==23.3.1
pydantic==2.5.2
pydantic-core==2.14.5
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools==68.2.0
sniffio==1.3.0
tqdm==4.66.1
typing-extensions==4.8.0
urllib3==2.1.0
wheel==0.41.2
yarl==1.9.3
selenium~=4.17.2
beautifulsoup4~=4.12.3